Add parsing to make multiple pipes possible
Added syntax parsing - currently just for the sake of piping.
The piping code does *not* handle builtins *at all*.
That will be saved for a later time.

An ASTNode represents a command. It can have a next_node, which will be
the next command in a pipeline.

With this, a major goal is accomplished, and the groundwork of more
complex functionaltiy is laid.