Add support for conditional exectuion
So-called short-circuiting with && and ||.

- && is handled in parse_statement, since it depends entirely on the
  *one* command before it
- || is handled in parse_line, since it depends on all prior commands
  (it has low priority, and if a command fails it should carry over)

None of these will mess up `;`,
and the line `true || ls ; ls` will run `ls` once.

Because of the way I wrote the parser, this required two more
run-throughs of the entire set of tokens. One for `;`, then one for `||`
and one for `&&` to handle the different priorities of the operators.