Turn exit into a builtin
Exit is now a builtin that returns NESSIE_EXIT_SUCCESS to indicate that
the user/program intends to exit. This value is checked for in
execute_syntax_tree and will always be returned if it is found.
All in all, this was a fairly straightforward change.

Since builtins are not able to be piped yet, there is no need to check
for `exit` inside a pipe. (and who would want to do that, anyway?)

Whenever that is changed, there needs to be some checks put in place
to ensure that this behavior is in line with POSIX shells.