Add more error handling
Just printing info with perror and exiting whenever an unrecoverable
error occurs - like a failed `malloc` or `pipe` call.

Added a utility function for doing just that.
CCLS complains about a null pointer dereference, that's blatantly wrong.
It does *not* complain if the `exit` call is written out.
However, I do not like the thought of having two lines where one is
needed just because the *linter* complains.