Proper string behaviour if close to words
If a string is adjacent to a word, it should be merged with the word.
The entire text should be treated as _one_ token.

An example:
$ echo this" is how it should "be
will print `this is how it should be` in POSIX shells,
and now also in Nessie.

This required just a few minor changes in the lexer.