Add support for variables (+ misc cleanup)
The support is in form of a "preprocessor" step that expands whatever
variables are present in the line to their known form.

Fixed some formatting issues and added a few debug things here and
there. Discovered some particularly odd behaviour that will have to be
investigated (builtin after && in some conditions, invalid pointer to
free etc.)