Fix regression caused by preprocessor code
In one specific test case, a mysterious final character appeared:

    echo $x; echo $y&& echo

where x and y were "correct" and "correct indeed."
This extra character (often ! or ASCII 17) prevented Nessie from
recognizing `echo` as a command.

The bug kept cropping up in all my attempts to crush it, until I finally
reached what seemed to be the cause. The `split_input` procedure ran
through *one more character* than it needed to, ending up with an
uninitialized value where there really should be a '\0'. This had to
mean that the preprocessor did not allocate *enough memory* for the last
character in the string.

It's all so simple now.

I ended up changing some other things in the process:

 - The lexer now takes a length argument, produced by the preprocessor
   This means that it no longer relies on there being a null byte at the
   end of the input, which might make it more robust to future mistakes.

 - In debug mode, each logged token is now displayed together with its
   size in bytes. This helped confirm that the problem was indeed a
   superfluous character popping up at the tail end of the expanded
   input.