Prevent long variables from crashing the shell
The length of the expanded text of a var didn't become long enough
(it couldn't even hold my PATH) and caused severe errors.
Fixed by adding a `max` resulting in `value_len` or `exp_len / 3`,
depending on which one is larger.

Made a test that goes up to 12800 chars – that should be plenty.